requirements.txt
python-dotenv
requests
pyTelegramBotAPI
fpdf
python-telegram-bot==13.15

